Toshiba

+++ src/com/company/Moto.java
@@ -1,7 +1,7 @@
 package com.company;
 
 public class Moto extends Vehiculo {
-    public Moto(Integer velocidad, Integer aceleracion, Integer anguloDeGiro, String patente) {
-        super(velocidad, aceleracion, anguloDeGiro, patente, 300, 2);
+    public Moto(Double velocidad, Double aceleracion, Double anguloDeGiro, String patente) {
+        super(velocidad, aceleracion, anguloDeGiro, patente, 300.0, 2);
     }
 }
